+++ react-components/src/architecture/concrete/primitives/line/LineDomainObject.ts
@@ -19,7 +19,7 @@ import { getIconByPrimitiveType } from '../../../base/utilities/primitives/getIc
 import { type TranslationInput } from '../../../base/utilities/TranslateInput';
 import { clear } from '../../../base/utilities/extensions/arrayExtensions';
 import { type IconName } from '../../../base/utilities/IconName';
-import { Vector3ArrayUtils } from '../../../base/utilities/primitives/PointsUtils';
+import { Vector3ArrayUtils } from '../../../base/utilities/primitives/Vector3ArrayUtils';
 import { DomainObjectTransaction } from '../../../base/undo/DomainObjectTransaction';
 import { type Transaction } from '../../../base/undo/Transaction';
